+++ __tests__/api/signup/register.test.ts
@@ -11,7 +11,7 @@ import {
 } from "@aws-sdk/client-cognito-identity-provider";
 import register from "@pages/api/signup/register";
 
-const mockGetCSRFToken = mocked(getCsrfToken, true);
+const mockGetCSRFToken = mocked(getCsrfToken, { shallow: true });
 
 jest.mock("next-auth/react");
 jest.mock("@aws-sdk/client-cognito-identity-provider", () => ({
@@ -66,8 +66,10 @@ describe("/signup/register", () => {
     });
   });
   describe("Sign Up Registration", () => {
-    const mockedCognitoIdentityProviderClient: any = mocked(CognitoIdentityProviderClient, true);
-    const mockedSignUpCommand: any = mocked(SignUpCommand, true);
+    const mockedCognitoIdentityProviderClient: any = mocked(CognitoIdentityProviderClient, {
+      shallow: true,
+    });
+    const mockedSignUpCommand: any = mocked(SignUpCommand, { shallow: true });
     const sendFunctionMock = jest.fn();
     afterEach(() => {
       mockedCognitoIdentityProviderClient.mockReset();