+++ components/form-builder/icons/index.ts
@@ -44,3 +44,6 @@ export { BulletListIcon } from "./BulletListIcon";
 export { LinkIcon } from "./LinkIcon";
 export { TreeViewIcon } from "./TreeViewIcon";
 export { RoundCloseIcon } from "./RoundCloseIcon";
+export { ShareExternalLinkIcon } from "./ShareExternalLinkIcon";
+export { CopyIcon } from "./CopyIcon";
+export { ChevronRight } from "./ChevronRight";