+++ CHANGELOG.md
@@ -28,6 +28,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - The existing `publishingStatus` field from the form JSON configuration has been replaced by a `isPublished` data field in the database. It can be switch to `true` or `false` using the Template API. A migration process will automatically happen through the Prisma seeding process. [#1181](https://github.com/cds-snc/platform-forms-client/issues/1181)
 - Form builder can only load form if the user has the permission to access it [#1228](https://github.com/cds-snc/platform-forms-client/issues/1228)
 - Form template JSON schema does not allow for addtional properties anymore. [#1455](https://github.com/cds-snc/platform-forms-client/issues/1455)
+- Reworked Template DB schema. [#1480](https://github.com/cds-snc/platform-forms-client/issues/1480)
 
 ### Fixed